hypervolume

+++ src/test/java/EdBarTest.java
@@ -38,15 +38,15 @@ public class EdBarTest extends JFrame {
     final XYSeriesCollection dataset = new XYSeriesCollection();
 
     for (int i = 0; i < ids.length; i++) {
-      Map<String, Map<Integer, Integer>> map = Indicators.quantityEdBySolutions(ids, ids[i]);
+      Map<String, Map<Double, Integer>> map = Indicators.quantityEdBySolutions(ids, ids[i]);
 
-      Map.Entry<String, Map<Integer, Integer>> content = map.entrySet().iterator().next();
+      Map.Entry<String, Map<Double, Integer>> content = map.entrySet().iterator().next();
       final XYSeries serie = new XYSeries(content.getKey());
 
-      Map<Integer, Integer> a = content.getValue();
+      Map<Double, Integer> a = content.getValue();
 
-      for (Map.Entry<Integer, Integer> entry : a.entrySet()) {
-        Integer double1 = entry.getKey();
+      for (Map.Entry<Double, Integer> entry : a.entrySet()) {
+        Double double1 = entry.getKey();
         Integer integer = entry.getValue();
         objDataset.addValue(integer, content.getKey(), double1);
       }
